Clean up main config menu
