+++ HRCounter/Installers/MenuInstaller.cs
@@ -7,7 +7,7 @@ public class MenuInstaller : Installer<MenuInstaller>
 {
     public override void InstallBindings()
     {
-        Container.BindInterfacesAndSelfTo<SettingMenuController>().FromNewComponentAsViewController().AsSingle();
+        Container.BindInterfacesAndSelfTo<MainConfigMenu>().FromNewComponentAsViewController().AsSingle();
         Container.BindInterfacesAndSelfTo<ServiceStatusMenu>().FromNewComponentAsViewController().AsSingle();
         Container.BindInterfacesAndSelfTo<DataSourceMenu>().FromNewComponentAsViewController().AsSingle();
         Container.BindInterfacesAndSelfTo<ConfigViewFlowCoordinator>().FromNewComponentOnNewGameObject().AsSingle().NonLazy();
