+++ HRCounter/UI/BSML/configMenu.bsml
@@ -18,17 +18,17 @@
         <vertical child-control-height='false' spacing="~vertical-spacing" horizontal-fit="PreferredSize" pref-width="~tab-width">
             <horizontal>
                 <vertical background="round-rect-panel" pad="1">
-                    <list id='icon-list' select-cell='icon-selected' show-scrollbar="false"/>
+                    <list id='icon-list' select-cell='OnIconListCellSelect' show-scrollbar="false"/>
                 </vertical>
                 <vertical child-control-height="false" pad-left="0" preferred-width="10" horizontal-fit="PreferredSize">
-                    <button text="Refresh" on-click="refresh-counter-icon" interactable="~is-icon-loaded" pad-top="4"/>
+                    <button text="Refresh" on-click="RefreshIconOptions" interactable="~IsIconListLoaded" pad-top="4"/>
                     <page-button event-click='icon-list#PageUp' direction='Up' size-delta-y="10" size-delta-x="10"/>
                     <page-button event-click='icon-list#PageDown' direction='Down' size-delta-y="10" size-delta-x="10"/>
                 </vertical>
             </horizontal>
             <horizontal horizontal-fit="PreferredSize" pref-width="100">
                 <text text="Put your icons in UserData/HRCounter/Icons/" align="Center" italics="false"/>
-                <button text="Show in Explorer" on-click="open-icons-folder" pad-left="8"/>
+                <button text="Show in Explorer" on-click="OpenIconsFolder" pad-left="8"/>
             </horizontal>
         </vertical>
     </tab>
@@ -42,21 +42,21 @@
                             apply-on-change="true" bind-value="true"/>
             <horizontal>
                 <color-setting text="Low Color" value="LowColor" apply-on-change="true" bind-value="true"/>
-                <button text="Reset" on-click="reset-low-color" pad-left="2"/>
+                <button text="Reset" on-click="ResetLowColor" pad-left="2"/>
             </horizontal>
             <horizontal>
                 <color-setting text="Middle Color" value="MidColor" apply-on-change="true" bind-value="true"/>
-                <button text="Reset" on-click="reset-mid-color" pad-left="2"/>
+                <button text="Reset" on-click="ResetMidColor" pad-left="2"/>
             </horizontal>
             <horizontal>
                 <color-setting text="High Color" value="HighColor" apply-on-change="true" bind-value="true"/>
-                <button text="Reset" on-click="reset-high-color" pad-left="2"/>
+                <button text="Reset" on-click="ResetHighColor" pad-left="2"/>
             </horizontal>
             <horizontal horizontal-fit="PreferredSize">
                 <horizontal horizontal-fit="PreferredSize" pref-width="50" pref-height="10" child-align="MiddleCenter">
-                    <text id="color-info-text" italics="false" align="Center"/>
+                    <text text="~ColorInfoText" italics="false" align="Center"/>
                 </horizontal>
-                <button text="~visualize-colors-btn-text" on-click="visualize-colors-btn-pressed" pad-left="2" pref-width="20"/>
+                <button text="~VisualizeColorsBtnText" on-click="VisualizeColorsBtnPressed" pad-left="2" pref-width="20"/>
             </horizontal>
         </vertical>
     </tab>
